touch trion to use the crocker.Dock standard for referencing a docker daemon.
thoughts on teh future: it would be nice to use something vaguely like a RAII pattern for docker container instances (creating the object is the same as launching a container), rather than the way we've been doing things so far, which is a bit ad-hoc and clearly admits to the docker daemon being the focal point instead of the containers.  But what we really want to talk about is the Containers.  The daemon is an incidental in their creation and management.  So, let's do that: one constructs a crocker.Container when one wants to launch something, and that happens to take an argument of a crocker.Dock; thereafter, there's never any reason to pass around or mention the crocker.Dock object.  This gets us closer to more sanely process-tree-oriented semantics (which, not coincidentally, is what lxc and cgroups look like internally, if we ever decide to wipe away the fog of docker entirely and do that lifting ourselves).