work on using temporary working trees for dex operations.
implementation allows passing a function to a withTempTree() function that sets up context and automatically tears it down after.

not quite all tests are rewritten to use it yet -- we're going to make the whole idea of PreparePublish() and similar methods go away before trying to fix tests that refer to them.