Correctly explain unix socket double-checking
The reason you need to double-check that a unix socket is dialable is that the daemon has no way to reliably clean it up. Whether docker attempts to clean it up some of the time is irrelevant - you should always double-check anyway.

If you would like to criticize or improve the behavior of Docker, please do it in a pull request or issue on http://github.com/dotcloud/docker, otherwise we don't get to improve Docker (or explain why it's really not a flaw after all).

Thanks.